Deprecating This Branch
Deprecating this branch due to not undestanding the scope of what I could do with a Jupyter Notebook.

I wanted to try making a notebook and then turning it into an executeable file with a cell containing a widget output as being the view but it is out of my depth.

Deprecating this branch to move into more standard Python development using a script and PyQT or Tkinter GUI support.