Testing UI Elements
Set up more of the Dev environment for designing the program. Installed PyQt tools to use Qt Designer, the GUI based UI design.

Also tested importing the raw UI file, as well as converting the UI file to a .py file to be imported into the project.

Tested how to do event handling and getting values for some widget elements.